Rename files to use dash
