+++ app-shell/src/app/fragments/base-fragment.ts
