+++ app-shell/src/app/lib/event-bus.ts
